requirements.txt
fastapi 
uvicorn[standard] 
PyYAML 
pydantic
websockets
requests
httpx[socks]
curl_cffi 
PySocks 
cloudflare
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
docker
pymysql
cryptography
cachetools